@@ -1,133 +1,43 @@
 import 'package:flutter/material.dart';
-import 'package:device_apps/device_apps.dart';
-// import 'package:cache_manager/cache_manager.dart';
-import 'package:android_intent_plus/android_intent.dart';
-import 'package:android_intent_plus/flag.dart';
-import 'package:app_settings/app_settings.dart';
-import 'package:nkit_launcher/app_list.dart';
-
-AppListCacher appList = AppListCacher();
-// FocusNode messageFocus = FocusNode();
 
 void main() {
   runApp(const MyApp());
 }
 
-class AppListWidget extends StatefulWidget {
-  const AppListWidget({Key? key, required this.filterKey}) : super(key: key);
-
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
-
-  final String filterKey;
-
-  @override
-  State<AppListWidget> createState() => AppListState();
-}
-
-class AppListState extends State<AppListWidget> {
-  Widget projectWidget() {
-    // messageFocus.requestFocus();
-    return FutureBuilder(
-      initialData: <App>[],
-      builder: (context, projectSnap) {
-        // messageFocus.requestFocus();
-        if (projectSnap.connectionState == ConnectionState.none &&
-            projectSnap.hasData == null) {
-          return Text('connection none');
-        }
-
-        if (projectSnap.data == null) {
-          return Text('data null');
-        }
-
-        List<App> apps = projectSnap.data as List<App>;
-
-        bool isFiltered = widget.filterKey != '';
-        apps.sort(
-          (a, b) => a.appName.toLowerCase().compareTo(b.appName.toLowerCase()),
-        );
-        if (isFiltered) {
-          apps = apps.reversed.toList();
-        }
-        apps.removeWhere((app) => !app.appName
-            .toLowerCase()
-            .contains(widget.filterKey.toLowerCase()));
-        print('App length: ${apps.length}');
-        return ListView.builder(
-          reverse: isFiltered,
-          itemCount: apps.length,
-          itemBuilder: (context, index) {
-            App app = apps[index];
-            return GestureDetector(
-              child: ListTile(
-                title:
-                    Text(app.appName, style: const TextStyle(fontSize: 18.0)),
-              ),
-              onTap: () => {DeviceApps.openApp(app.packageName)},
-              onLongPress: () {
-                AndroidIntent intent = AndroidIntent(
-                  action: 'action_application_details_settings',
-                  data: 'package:${app.packageName}',
-                  flags: <int>[Flag.FLAG_ACTIVITY_NEW_TASK],
-                );
-                intent.launch();
-              },
-            );
-          },
-        );
-      },
-      future: appList.getAppList(),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: projectWidget(),
-    );
-  }
-}
-
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Nkit Launcher',
+      title: 'Flutter Demo',
       theme: ThemeData(
-        brightness: Brightness.light
+        // This is the theme of your application.
+        //
+        // TRY THIS: Try running your application with "flutter run". You'll see
+        // the application has a purple toolbar. Then, without quitting the app,
+        // try changing the seedColor in the colorScheme below to Colors.green
+        // and then invoke "hot reload" (save your changes or press the "hot
+        // reload" button in a Flutter-supported IDE, or press "r" if you used
+        // the command line to start the app).
+        //
+        // Notice that the counter didn't reset back to zero; the application
+        // state is not lost during the reload. To reset the state, use hot
+        // restart instead.
+        //
+        // This works for code too, not just values: Most code changes can be
+        // tested with just a hot reload.
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepOrange),
+        useMaterial3: true,
       ),
-      darkTheme: ThemeData(
-          // backgroundColor: const Color(0xff000000)
-          brightness: Brightness.dark),
-      // theme: ThemeData(
-      //   // This is the theme of your application.
-      //   //
-      //   // Try running your application with "flutter run". You'll see the
-      //   // application has a blue toolbar. Then, without quitting the app, try
-      //   // changing the primarySwatch below to Colors.green and then invoke
-      //   // "hot reload" (press "r" in the console where you ran "flutter run",
-      //   // or simply save your changes to "hot reload" in a Flutter IDE).
-      //   // Notice that the counter didn't reset back to zero; the application
-      //   // is not restarted.
-      //   primarySwatch: Colors.blue,
-      // ),
-      home: const MyHomePage(title: 'Nkit Launcher Home Page'),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  const MyHomePage({Key? key, required this.title}) : super(key: key);
+  const MyHomePage({super.key, required this.title});
 
   // This widget is the home page of your application. It is stateful, meaning
   // that it has a State object (defined below) that contains fields that affect
@@ -145,31 +55,71 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  String filterSearch = '';
+  int _counter = 0;
+
+  void _incrementCounter() {
+    setState(() {
+      // This call to setState tells the Flutter framework that something has
+      // changed in this State, which causes it to rerun the build method below
+      // so that the display can reflect the updated values. If we changed
+      // _counter without calling setState(), then the build method would not be
+      // called again, and so nothing would appear to happen.
+      _counter++;
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
+    // This method is rerun every time setState is called, for instance as done
+    // by the _incrementCounter method above.
+    //
+    // The Flutter framework has been optimized to make rerunning build methods
+    // fast, so that you can just rebuild anything that needs updating rather
+    // than having to individually change instances of widgets.
     return Scaffold(
-        body: Column(
-          children: [
-            Expanded(
-              child: AppListWidget(
-                filterKey: filterSearch,
-              ),
+      appBar: AppBar(
+        // TRY THIS: Try changing the color here to a specific color (to
+        // Colors.amber, perhaps?) and trigger a hot reload to see the AppBar
+        // change color while the other colors stay the same.
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        // Here we take the value from the MyHomePage object that was created by
+        // the App.build method, and use it to set our appbar title.
+        title: Text(widget.title),
+      ),
+      body: Center(
+        // Center is a layout widget. It takes a single child and positions it
+        // in the middle of the parent.
+        child: Column(
+          // Column is also a layout widget. It takes a list of children and
+          // arranges them vertically. By default, it sizes itself to fit its
+          // children horizontally, and tries to be as tall as its parent.
+          //
+          // Column has various properties to control how it sizes itself and
+          // how it positions its children. Here we use mainAxisAlignment to
+          // center the children vertically; the main axis here is the vertical
+          // axis because Columns are vertical (the cross axis would be
+          // horizontal).
+          //
+          // TRY THIS: Invoke "debug painting" (choose the "Toggle Debug Paint"
+          // action in the IDE, or press "p" in the console), to see the
+          // wireframe for each widget.
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text(
+              'You have pushed the button this many times:',
             ),
-            TextField(
-              decoration: const InputDecoration(
-                border: UnderlineInputBorder(),
-                contentPadding: EdgeInsets.all(8.0),
-                hintText: 'Search',
-              ),
-              onChanged: (value) => {
-                setState(() => {filterSearch = value})
-              },
-              autocorrect: false,
-              autofocus: true,
-              // focusNode: messageFocus,
+            Text(
+              '$_counter',
+              style: Theme.of(context).textTheme.headlineMedium,
             ),
           ],
-        ));
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: _incrementCounter,
+        tooltip: 'Increment',
+        child: const Icon(Icons.add),
+      ), // This trailing comma makes auto-formatting nicer for build methods.
+    );
   }
 }
